+++ earth2studio/data/cbottle.py
@@ -1,4 +1,4 @@
-# SPDX-FileCopyrightText: Copyright (c) 2024-2025 NVIDIA CORPORATION & AFFILIATES.
+# SPDX-FileCopyrightText: Copyright (c) 2024-2026 NVIDIA CORPORATION & AFFILIATES.
 # SPDX-FileCopyrightText: All rights reserved.
 # SPDX-License-Identifier: Apache-2.0
 #
@@ -225,7 +225,8 @@ class CBottle3D(torch.nn.Module, AutoModelMixin):
             batch["second_of_day"] = second_of_day[start_idx:end_idx]
             batch["day_of_year"] = day_of_year[start_idx:end_idx]
 
-            output, coords = self.core_model.sample(batch, seed=self.seed)
+            seed = None if self.seed is None else self.seed + i
+            output, coords = self.core_model.sample(batch, seed=seed)
             output = output[:, varidx]
             outputs.append(output)
 
@@ -370,7 +371,7 @@ class CBottle3D(torch.nn.Module, AutoModelMixin):
     def load_default_package(cls) -> Package:
         """Default pre-trained CBottle3D model package from Nvidia model registry"""
         return Package(
-            "ngc://models/nvidia/earth-2/cbottle@1.2",
+            "hf://nvidia/cbottle@eebd93c85b3cd3a5a8f79c546ed917b0b80438f4",
             cache_options={
                 "cache_storage": Package.default_cache("cbottle"),
                 "same_names": True,
@@ -428,6 +429,11 @@ class CBottle3D(torch.nn.Module, AutoModelMixin):
         core_model = MixtureOfExpertsDenoiser.from_pretrained(
             checkpoints, (100.0, 10.0)
         )
+
+        try:
+            package.resolve("config.json")  # HF tracking download statistics
+        except FileNotFoundError:
+            pass
 
         # The following code is left here for reference of how to access the AMIP SST
         # data from the original data store. NGC is faster and cleaner so it is also